Converting Image Data to pytorch tensor
This is an update to Issue 172. 

Currently, I'm working on finding a way to convert the image data into a Pytorch tensor to be used with Pytorch models. 
The issue that I've been running into is that the current method of loading the image using the Spectral library makes a Numpy memmap that isn't compatible with Pytorch tensors. I'm looking into the Spectral documentation to find a way to possibly make conversion from this Numpy memmap to Pytorch tensors possible.

I will be updating this branch as I go along.